Remove file change detection
Turns out that there's no reliable way of figuring out whether a plaintext
file is no longer being edited in a text editor, so don't bother with it.
Instead, display a messagebox asking the user whether they want to encrypt
or discard their changes as soon as the file is opened, allowing them to
click yes or no when they're finished.

It's not exactly pretty, but it's better than running the risk of leaving
plaintext files in the password store without the user knowing.