Reorder initialisation code
App initialisation is now done in a more sensible order.
First, a notification area icon is created, so any notifications
created during initialisation don't just vanish into nothingness.

Next, the configuration file is loaded. Error handling for this has been
improved as well.
After that, hotkeys are assigned, the main window name is set, and the
GPG agent is preloaded.