Use a better method to preload the GPG agent
The previous method worked, but caused GPG to produce an error, which
got turned into an exception that had to be caught.

There's a much better way to start the GPG agent, which is to simply
asks GPG for a list of its private keys.

By using this method, we can get rid of the exception handler while
making the code a little clearer as well.