Switch back to calling gpg.exe directly
Turns out gpgme.dll is, unfortunately, not very reliable. It won't work
if you simply bundle GPG with your application, only if you actually run
the installer for GPG. It also sometimes causes issues with decryption
when the agent hasn't yet stored the password. Because of this and for
several other reasons, we're going back to the old method of calling
gpg.exe directly (which is actually how gpgme does it as well).