Selection window: Move drawing code out of OnContentRendered
By creating a transformation matrix from a HwndSource, we can get the
DPI settings for this window immediately, instead of having to wait for
the window to be rendered so its CompositionTarget becomes available.
This means we can move the size/location setup out of OnContentRendered,
so our window instantly gains the right size and location, instead of
jumping to the right position after first being rendered in the top
left of the screen.